+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
   - https://github.com/Shopify/flash-list/pull/518
 - Upgrade recyclerlistview to v4.1.1
   - https://github.com/Shopify/flash-list/pull/526
-- Add Known Issue for useState in cell components
+- Add common pitfall for useState in recycled components
   - https://github.com/Shopify/flash-list/pull/529
 
 ## [1.1.0] - 2022-07-06