+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Fixed out of bound read from data
+  - https://github.com/Shopify/flash-list/pull/523
 
 - Added JS only fallbacks for unsupported platforms
   - https://github.com/Shopify/flash-list/pull/518